requirements_abstractive.txt
transfopytorch_lightning
absl-py
rouge_score
nltk
requestsorch
sentencepiece
protobuf
